\name{RKerf-package}
\alias{RKerf-package}
\alias{RKerf}
\docType{package}
\title{
\packageTitle{RKerf}
}
\description{
\packageDescription{RKerf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RKerf}
\packageIndices{RKerf}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RKerf}

Maintainer: \packageMaintainer{RKerf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
